Fix version number and man page building.
The 3 was a build number not a dot release. This is fixed in upstream tagging.

The manpage in the original package is compiled by ronn, which PLD doesn't have. The upstream project maintains a patch with a precompiled version of the manpage for some versions of Debian that also lack ronn. An upstream branch for PLD has been created with these fixes and other tweaks.